requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-yasg==1.21.7
pandas==2.1.1
openpyxl==3.1.2
Pillow==10.0.1
django-filter==23.5
python-dateutil==2.8.2
